criando classes de dominio para especialidades, clubes, requisitos, classes etc

+++ src/main/java/br/com/serrano/service/AreaService.java
@@ -4,6 +4,7 @@ import br.com.serrano.model.Area;
 import br.com.serrano.repository.AreaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 @Service
 public class AreaService {
@@ -14,12 +15,17 @@ public class AreaService {
     }
     public Area getAreaById(Integer id){ return areaRepository.getAreaById(id);}
     public Area createOrUpdateArea(Area area){return areaRepository.save(area);}
-    public String deleteUnidade(Integer id){
+    public String deleteArea(Integer id){
+        Area area = areaRepository.getAreaById(id);
         try {
-            areaRepository.deleteById(id);
-            return "Area deletado com sucesso";
+            if(area != null){
+                areaRepository.deleteById(id);
+                return "Area deletado com sucesso";
+            } else {
+                return "Area nao encontrada na base de dados";
+            }
         }catch(Exception e) {
-            return "Erro ao deletar o area";
+            return "Erro ao deletar a area";
         }
     }
 }
