+++ src/main/java/br/com/serrano/service/EspecialidadeService.java
@@ -4,6 +4,7 @@ import br.com.serrano.model.Especialidade;
 import br.com.serrano.repository.EspecialidadeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 @Service
 public class EspecialidadeService {
@@ -15,9 +16,14 @@ public class EspecialidadeService {
     public Especialidade getEspecialidadeById(Integer id){ return especialidadeRepository.getEspecialidadeById(id);}
     public Especialidade createOrUpdateEspecialidade(Especialidade unidade){return especialidadeRepository.save(unidade);}
     public String deleteEspecialidade(Integer id){
+        Especialidade especialidade = especialidadeRepository.getEspecialidadeById(id);
         try {
-            especialidadeRepository.deleteById(id);
-            return "Especialidade deletado com sucesso";
+            if(especialidade != null){
+                especialidadeRepository.deleteById(id);
+                return "Especialidade deletado com sucesso";
+            } else {
+                return "Especialidade nao encontrada na base de dados";
+            }
         }catch(Exception e) {
             return "Erro ao deletar o especialidade";
         }
