+++ src/main/java/br/com/serrano/service/RequisitoService.java
@@ -16,9 +16,14 @@ public class RequisitoService {
     public Requisito getRequisitoById(Integer id){ return requisitoRepository.getRequisitoById(id);}
     public Requisito createOrUpdateRequisito(Requisito requisito){return requisitoRepository.save(requisito);}
     public String deleteRequisito(Integer id){
+        Requisito requisito = requisitoRepository.getRequisitoById(id);
         try {
-            requisitoRepository.deleteById(id);
-            return "Requisito deletado com sucesso";
+            if(requisito != null){
+                requisitoRepository.deleteById(id);
+                return "Requisito deletado com sucesso";
+            } else {
+                return "Requisito nao encontrado na base de dados";
+            }
         }catch(Exception e) {
             return "Erro ao deletar o requisito";
         }
