+++ src/main/java/br/com/serrano/service/UnidadeService.java
@@ -4,6 +4,7 @@ import br.com.serrano.model.Unidade;
 import br.com.serrano.repository.UnidadeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 @Service
 public class UnidadeService {
@@ -14,11 +15,15 @@ public class UnidadeService {
     }
     public Unidade getUnidadeById(Integer id){ return unidadeRepository.getUnidadeById(id);}
     public Unidade createOrUpdateUnidade(Unidade unidade){return unidadeRepository.save(unidade);}
-
     public String deleteUnidade(Integer id){
+        Unidade unidade = unidadeRepository.getUnidadeById(id);
         try {
-            unidadeRepository.deleteById(id);
-            return "Unidade deletado com sucesso";
+            if(unidade != null){
+                unidadeRepository.deleteById(id);
+                return "Unidade deletado com sucesso";
+            } else {
+                return "Unidade nao encontrado na base de dados";
+            }
         }catch(Exception e) {
             return "Erro ao deletar o unidade";
         }
